% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estTrialInfectRate.R
\name{calcParam}
\alias{calcParam}
\title{Calculate the key parameters for ESCUDDO}
\usage{
calcParam(
  pSA = c(0, 0, 0.1, 0.26, 0.39, 0.53, 0.66, 0.77, 0.84, 0.91, 0.95),
  pEA = c(0.25, 0.2, 0.19, 0.19, 0.17),
  hI1 = c(0, 0, 0.03, 0.04, 0.08, 0.1, 0.1, 0.1, 0.1),
  hI2 = c(0, 0, 0.03, 0.03, 0.05, 0.07, 0.07, 0.07, 0.07),
  npop = 5e+05,
  ofile = NA
)
}
\arguments{
\item{pSA}{A vector of 11 values. Probability of starting sexual activity by ages 12:22}

\item{pEA}{A vector of 5 values.  Proportion of subjects who enroll at ages 12:16}

\item{hI1}{A vector of 9 values.  Herd Immunity for individuals in 2018:2026 (<= 18 yo )}

\item{hI2}{A vector of 9 values.  Herd Immunity for individuals in 2018:2026 (>  18 yo )}

\item{npop}{A single value. Total number of women in the simulated population}

\item{ofile}{A string. The file name of where to store key parameters for simulation.}
}
\value{
p0   A value. The proportion of women who have an infection at baseline

pi0.amongBaselineUnifected.4yr A value. Among baseline (0/6 month) uninfected women, this value  is the proportion who have an incident persistent infection in four years

pi0.amongBaselineUnifected.5yr A value. Among baseline (0/6 month) uninfected women, this value  is the proportion who have an incident persistent infection in five years

delta0.4yr A value. Among baseline-0 (only 0 month) uninfected women, this value  is the proportion who have a persistent infection spanning 3.5 and 4.0  years

delta0.5yr A value. Among baseline-0 (only 0 month) uninfected women, this value  is the proportion who have a persistent infection spanning 4.5 and 5.0  years
}
\description{
Calculate the key parameters for ESCUDDO
}
\examples{
myRes <- calcParam(
hI1     = rep(0,9),
hI2     = rep(0,9),
npop    = 5000000,
ofile   = "/volumes/data/Projects/HPV/out/Rparams.Rdat")
}
