% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estTrialInfectRate.R
\name{calcProbNewInf}
\alias{calcProbNewInf}
\title{Calculate the probability of acquiring a new infection in 6 months}
\usage{
calcProbNewInf()
}
\value{
The estimated probability of acquiring a new infection 6 months after the last visit
}
\description{
Calculate the probability of acquiring a new infection in 6 months
}
