% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estVE.R
\name{confInt}
\alias{confInt}
\title{Obtain the confidence interval for the difference in two rates (method 3: C.P. Farrington and G. Manning
Stat. Med., 9 (1990), pp. 1447-1454).}
\usage{
confInt(n1, N1, n2, N2, c = 0.95)
}
\arguments{
\item{n1}{A single value. Number of events in the first group}

\item{N1}{A single value. Number of individuals in the first group}

\item{n2}{A single value. Number of events in the second group}

\item{N2}{A single value. Number of individuals in the second group}

\item{c}{A single value. A confidence level (default c = 0.95)}
}
\value{
L     A single value. Lower Bound

U     A single value. Upper Bound

D     A single value. The difference in rates (n1/N1-n2/N2)
}
\description{
Obtain the confidence interval for the difference in two rates
}
\examples{
confInt(5,100,4,100)

}
