% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estVE.R
\name{eventCount}
\alias{eventCount}
\title{Obtain the observed and expected number of events in one arm of the trial}
\usage{
eventCount(dMat16x, dMat18x, sMatx, ATPx)
}
\arguments{
\item{dMat16x}{A matrix (nrow=number.of.subjects, ncol=number.of.visits.including.baseline). Entries are 0 or 1, indicating HPV 16 infection status (w/ -9 for missing)}

\item{dMat18x}{A matrix (nrow=number.of.subjects, ncol=number.of.visits.including.baseline). Entries are 0 or 1, indicating HPV 18 infection status (w/ -9 for missing)}

\item{sMatx:}{A matrix (nrow=number.of.subjects, ncol=number.of.visits.including.baseline). Entries are 0 or 1, indicating if a woman has had her sexual debut (w/ -9 for missing)}

\item{ATPx:}{A vector of length number.of.subjects. Entries are 0 or 1, indicating whether the subject was part of the ATP group.}
}
\value{
N:      A single value. The number of women in the ATP cohort.

nobs:   A single value. The number of ATP women OBSERVED to have an event  (i.e. no 0/6 month infection and a incident persistent infection)

nexp:   A single value. The number of ATP women EXPECTED to have an event  (i.e. no 0/6 month infection and a incident persistent infection)

infFact:   A single value. The inflation factor = nexp/nobs

Nequiv:    A single value. The equivalent sample size = N/infFact
}
\description{
Obtain the observed and expected number of events in one arm of the trial
}
\examples{
trialRes <- simTrialArm(ofile="/volumes/data/Projects/HPV/out/Rparams.Rdat",yr=5,nsub=5000,pdis=0.01,pdo6=0.05,pdo=0.02,pmv=0.17)
dMat16x   <- trialRes$dMat16
dMat18x   <- trialRes$dMat18
sMatx     <- trialRes$sMat
ATPx      <- trialRes$ATPx
ecount    <- eventCount(dMat16x,dMat18x,sMatx,ATPx)
}
