% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estVE.R
\name{probInf}
\alias{probInf}
\title{Estimate the Probability that a participant had both (i) no baseline (0/6 month) infection (ii) an incident/persistent HPV-16 (or HPV-18) Infection}
\usage{
probInf(dMat16x, sMatx, ATPx)
}
\arguments{
\item{dMat16x}{A matrix (nrow=number.of.subjects, ncol=number.of.visits). Entries are HPV 16 (or HPV 18) infection status (w/ -9 for missing)}

\item{sMatx:}{A matrix (nrow=number.of.subjects, ncol=number.of.visits). Entries are of sexual activity status (w/ -9 for missing)}

\item{ATPx:}{A vector of length number.of.subjects. Entries show whether the subject was part of the ATP group.}
}
\value{
infecOUT. A matrix (nrow=length number.of.subjects, ncol = 2). The first column indicates if an event was observed. The second column shows the probability that
an event occurred. An event is two baseline negative results (i.e. 0 and 6 months) and an incident prevalent infection (two consecutive positive visits -- no intervening
negative or missing visits). non-ATP individuals are set to missing. The second column is based on the methods described in Sampson, Gail, CCT.
}
\description{
Estimate the Probability that a participant had both (i) no baseline (0/6 month) infection (ii) an incident/persistent HPV-16 (or HPV-18) Infection
}
\examples{
trialRes <- simTrialArm(ofile="/volumes/data/Projects/HPV/out/Rparams.Rdat",yr=5,nsub=5000,pdis=0.01,pdo6=0.05,pdo=0.02,pmv=0.17)
dMat16x   <- trialRes$dMat
sMatx     <- trialRes$sMat
ATPx      <- trialRes$ATPx
probInf2   <- probInf(dMat16x,sMatx,ATPx)
}
