% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estVE.R
\name{simStudy}
\alias{simStudy}
\title{Combine survey and trial participants}
\usage{
simStudy(trialRes, surveyRes)
}
\arguments{
\item{trialRes}{An object. The output from simTrialArm.}

\item{surveyRes}{An object. The output from simSurvey}
}
\value{
data      A matrix. The columns are GROUP (0=survey, 1=trial); DIS1.16 and DIS2.16 (HPV 16 status at 2 key visits);
DIS1.18 and DIS2.18 (HPV 18 status at 2 key visits); DIS0.16 (HPV 16 status at baseline; missing for survey);
DIS0.18 (HPV18 status at baseline; missing for survey); SEX1 and SEX2 (sexual activity status at two key visits);
X1 and X2 (two covariates); ATP (ATP status for trial participants; All survey participants have a value of 1);
}
\description{
Combine survey and trial participants
}
\examples{
trial1dose <- simTrialArm(ofile="/volumes/data/Projects/HPV/out/Rparams.Rdat",yr=5,nsub=5000,pdisL2=0.04*(1-0.8),pdo6=0.05,pdo=0.02,pmv=0.17)
survey     <- simSurvey(ofile="/volumes/data/Projects/HPV/out/Rparams.Rdat",  yr=5,nsub=5000,pdisL2=0.04,pmv2=0.17,pmv1=0.10)
study      <- simStudy(trialRes=trial1dose,surveyRes=survey)
}
