% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estVE.R
\name{simTrialArm}
\alias{simTrialArm}
\title{Generate a set of trial participants}
\usage{
simTrialArm(
  ofile = NA,
  yr = 5,
  nsub = 5000,
  pdis = NA,
  pdisL2 = NA,
  pdo6 = 0.05,
  pdo = 0.02,
  pmv = 0.17,
  seed = NA
)
}
\arguments{
\item{ofile}{A string. The file name for the file containing the key parameters (generated by calcParam()).}

\item{yr}{A single value. The number of years of the trial (either  4 or 5)}

\item{nsub}{A single value. The number of subjects the arm.}

\item{pdis}{A single value. The risk of being baseline uninfected and acquiring a new 16 or 18 persistent infection.}

\item{pdisL2}{A single value. The risk of being baseline uninfected and having a HPV 16 or 18 persistent infection at the last two visits
(NOTE: only pdis or pdisL2 should be set to a non-missing value)}

\item{pdo6}{A single value. The probability of dropping out before the 6 month visit}

\item{pdo}{A single value. The probability of dropping out before any other follow-up visit}

\item{pmv}{A single value. The probability of missing a visit (or specimen failure) at all follow-up vists after 6 months}
}
\value{
dMat:      A matrix (nrow=number.of.subjects, ncol=number.of.visits). Entries are infection HPV16/18 status (w/ -9 for missing)

dMat16:    A matrix (nrow=number.of.subjects, ncol=number.of.visits). Entries are HPV 16 infection status (w/ -9 for missing)

dMat18:    A matrix (nrow=number.of.subjects, ncol=number.of.visits). Entries are HPV 18 infection status (w/ -9 for missing)

sMat:      A matrix (nrow=number.of.subjects, ncol=number.of.visits). Entries are of sexual activity status (no missing).

infecTrue: A vecor of length number.of.subjects. Entries show whether the subject was both baseline negative and had a persistent infection.

ATPx:      A vecor of length number.of.subjects. Entries show whether the subject received both vaccinations.
}
\description{
Generate a set of trial participants
}
\examples{
trialRes <- simTrialArm(ofile="/volumes/data/Projects/HPV/out/Rparams.Rdat",yr=5,nsub=5000,pdis=0.01,pdo6=0.05,pdo=0.02,pmv=0.17)
}
